requirements.txt
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
PyJWT==1.7.1
python-dotenv==0.15.0
pytz==2019.3
requests==2.23.0
six==1.14.0
twilio==6.36.0
urllib3==1.25.8
Werkzeug==1.0.0
